+++ packages/svelte/src/internal/client/dom/blocks/svelte-element.js
@@ -12,30 +12,8 @@ import { set_should_intro } from '../../render.js';
 import { current_each_item, set_current_each_item } from './each.js';
 import { current_component_context, current_effect } from '../../runtime.js';
 import { DEV } from 'esm-env';
-import { is_array } from '../../utils.js';
-import { push_template_node } from '../template.js';
+import { assign_nodes } from '../template.js';
 import { noop } from '../../../shared/utils.js';
-
-/**
- * @param {import('#client').Effect} effect
- * @param {Element} from
- * @param {Element} to
- * @returns {void}
- */
-function swap_block_dom(effect, from, to) {
-	const dom = effect.dom;
-
-	if (is_array(dom)) {
-		for (let i = 0; i < dom.length; i++) {
-			if (dom[i] === from) {
-				dom[i] = to;
-				break;
-			}
-		}
-	} else if (dom === from) {
-		effect.dom = to;
-	}
-}
 
 /**
  * @param {Comment | Element} node
@@ -63,18 +41,6 @@ export function element(node, get_tag, is_svg, render_fn, get_namespace, locatio
 	/** @type {import('#client').Effect | null} */
 	let effect;
 
-	const parent_effect = /** @type {import('#client').Effect} */ (current_effect);
-
-	// Remove the the hydrated effect dom entry for our dynamic element
-	if (hydrating && is_array(parent_effect.dom)) {
-		var remove_index = parent_effect.dom.indexOf(
-			/** @type {import('#client').TemplateNode} */ (element)
-		);
-		if (remove_index !== -1) {
-			parent_effect.dom.splice(remove_index, 1);
-		}
-	}
-
 	/**
 	 * The keyed `{#each ...}` item block, if any, that this element is inside.
 	 * We track this so we can set it when changing the element, allowing any
@@ -82,8 +48,7 @@ export function element(node, get_tag, is_svg, render_fn, get_namespace, locatio
 	 */
 	let each_item_block = current_each_item;
 
-	block(() => {
-		const element_effect = /** @type {import('#client').Effect} */ (current_effect);
+	block(anchor, 0, () => {
 		const next_tag = get_tag() || null;
 		const ns = get_namespace
 			? get_namespace()
@@ -125,6 +90,8 @@ export function element(node, get_tag, is_svg, render_fn, get_namespace, locatio
 						? document.createElementNS(ns, next_tag)
 						: document.createElement(next_tag);
 
+				assign_nodes(element, element);
+
 				if (DEV && location) {
 					// @ts-expect-error
 					element.__svelte_meta = {
@@ -137,10 +104,7 @@ export function element(node, get_tag, is_svg, render_fn, get_namespace, locatio
 				}
 
 				if (prev_element && !hydrating) {
-					swap_block_dom(element_effect, prev_element, element);
 					prev_element.remove();
-				} else {
-					push_template_node(element, element_effect);
 				}
 
 				if (render_fn) {