+++ packages/svelte/src/internal/client/reactivity/effects.js
@@ -31,10 +31,10 @@ import {
 	DERIVED,
 	UNOWNED,
 	CLEAN,
-	INSPECT_EFFECT
+	INSPECT_EFFECT,
+	HEAD_EFFECT
 } from '../constants.js';
 import { set } from './sources.js';
-import { remove } from '../dom/reconciler.js';
 import * as e from '../errors.js';
 import { DEV } from 'esm-env';
 import { define_property } from '../utils.js';
@@ -75,16 +75,17 @@ export function push_effect(effect, parent_effect) {
  * @param {number} type
  * @param {null | (() => void | (() => void))} fn
  * @param {boolean} sync
+ * @param {boolean} push
  * @returns {import('#client').Effect}
  */
-function create_effect(type, fn, sync) {
+function create_effect(type, fn, sync, push = true) {
 	var is_root = (type & ROOT_EFFECT) !== 0;
 
 	/** @type {import('#client').Effect} */
 	var effect = {
 		ctx: current_component_context,
 		deps: null,
-		dom: null,
+		nodes: null,
 		f: type | DIRTY,
 		first: null,
 		fn,
@@ -120,10 +121,10 @@ function create_effect(type, fn, sync) {
 		sync &&
 		effect.deps === null &&
 		effect.first === null &&
-		effect.dom === null &&
+		effect.nodes === null &&
 		effect.teardown === null;
 
-	if (!inert && !is_root) {
+	if (!inert && !is_root && push) {
 		if (current_effect !== null) {
 			push_effect(effect, current_effect);
 		}
@@ -298,16 +299,22 @@ export function template_effect(fn) {
 }
 
 /**
- * @param {(() => void)} fn
+ * @param {import('#client').TemplateNode | null} anchor
  * @param {number} flags
+ * @param {(() => void)} fn
  */
-export function block(fn, flags = 0) {
-	return create_effect(RENDER_EFFECT | BLOCK_EFFECT | flags, fn, true);
+export function block(anchor, flags, fn) {
+	const effect = create_effect(RENDER_EFFECT | BLOCK_EFFECT | flags, fn, true);
+	if (anchor !== null) effect.nodes = { start: null, anchor: null, end: anchor };
+	return effect;
 }
 
-/** @param {(() => void)} fn */
-export function branch(fn) {
-	return create_effect(RENDER_EFFECT | BRANCH_EFFECT, fn, true);
+/**
+ * @param {(() => void)} fn
+ * @param {boolean} [push]
+ */
+export function branch(fn, push = true) {
+	return create_effect(RENDER_EFFECT | BRANCH_EFFECT, fn, true, push);
 }
 
 /**
@@ -335,13 +342,26 @@ export function execute_effect_teardown(effect) {
  * @returns {void}
  */
 export function destroy_effect(effect, remove_dom = true) {
-	var dom = effect.dom;
+	var removed = false;
 
-	if (dom !== null && remove_dom) {
-		remove(dom);
+	if ((remove_dom || (effect.f & HEAD_EFFECT) !== 0) && effect.nodes !== null) {
+		/** @type {import('#client').TemplateNode | null} */
+		var node = get_first_node(effect);
+		var end = effect.nodes.end;
+
+		while (node !== null) {
+			/** @type {import('#client').TemplateNode | null} */
+			var next =
+				node === end ? null : /** @type {import('#client').TemplateNode} */ (node.nextSibling);
+
+			node.remove();
+			node = next;
+		}
+
+		removed = true;
 	}
 
-	destroy_effect_children(effect, remove_dom);
+	destroy_effect_children(effect, remove_dom && !removed);
 	remove_reactions(effect, 0);
 	set_signal_status(effect, DESTROYED);
 
@@ -365,11 +385,42 @@ export function destroy_effect(effect, remove_dom = true) {
 		effect.prev =
 		effect.teardown =
 		effect.ctx =
-		effect.dom =
 		effect.deps =
 		effect.parent =
 		effect.fn =
+		effect.nodes =
 			null;
+}
+
+/**
+ * @param {import('#client').Effect} effect
+ * @returns {import('#client').TemplateNode}
+ */
+export function get_first_node(effect) {
+	var nodes = /** @type {NonNullable<typeof effect.nodes>} */ (effect.nodes);
+	var start = nodes.start;
+
+	if (start === undefined) {
+		// edge case — a snippet or component was the first item inside the effect,
+		// but it didn't render any DOM. in this case, we return the item's anchor
+		return /** @type {import('#client').TemplateNode} */ (nodes.anchor);
+	}
+
+	if (start !== null) {
+		return start;
+	}
+
+	var child = effect.first;
+	while (child && (child.nodes === null || (child.f & HEAD_EFFECT) !== 0)) {
+		child = child.next;
+	}
+
+	if (child !== null && child.nodes !== null) {
+		return get_first_node(child);
+	}
+
+	// in the case that there's no DOM, return the first anchor
+	return nodes.end;
 }
 
 /**