+++ packages/svelte/src/internal/client/types.d.ts
@@ -71,7 +71,7 @@ export type EachState = {
 	/** a key -> item lookup */
 	items: Map<any, EachItem>;
 	/** head of the linked list of items */
-	next: EachItem | null;
+	first: EachItem | null;
 };
 
 export type EachItem = {
@@ -85,7 +85,7 @@ export type EachItem = {
 	i: number | Source<number>;
 	/** key */
 	k: unknown;
-	prev: EachItem | EachState;
+	prev: EachItem | null;
 	next: EachItem | null;
 };