+++ packages/svelte/tests/compiler-errors/samples/runes-props-illegal-name-1/main.svelte
@@ -1,0 +1,3 @@
+<script>
+	let { $$slots: a } = $props();
+</script>