+++ packages/svelte/tests/runtime-browser/samples/bind-volume/main.svelte
@@ -1,0 +1,6 @@
+<script>
+	let volume = $state(0.1);
+</script>
+
+<audio bind:volume></audio>
+<button onclick={() => (volume += 0.1)}>increment</button>